requirements.txt
python-telegram-bot>=20.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
python-dotenv>=1.0.0
